kitchens: root-cause fix — /sync/push was skipping auto-share fanout
Native-app recipe creates flow through POST /api/sync/push, not
POST /api/recipes. The old _autoShareNewRecipe helper was defined
inline in recipes.js and only called from the REST create paths, so
every recipe an Android user created got INSERTed on the server
without ever triggering the Kitchen fan-out. Members never saw
those recipes and the auto_shared_count stayed 0 for anything
created on mobile - exactly the symptoms reported.

Fix:
- Extracted the helper to server/lib/auto-share.js so its home is
  shared between the two insert paths that exist today (and any
  future paths).
- sync.js now calls autoShareNewRecipe(u, serverId) after every
  recipes-table insert in POST /sync/push. try/catch around it so a
  fan-out hiccup cant kill the sync response.
- recipes.js imports the shared helper; local copy deleted.
- Logging left in place (per-run info line) so future misfires are
  visible in server logs.

The 78afc0b Resync button + count fix stay in place; Resync is
still the recovery path for the existing bad state on your db.
Going forward every insert path is covered.
